feat: PI screen full overhaul + SecureStorage crash fix
PI screen:
- Split ColonySummary storage into sfCapacityM3/sfUsedM3 (Storage Facility,
  12k m³ P0 buffer) and lpCapacityM3/lpUsedM3 (Launchpad, 10k m³ exports)
- Show two separate progress bars per colony — SF fill = P0 buffer state,
  LP fill = finished-goods readiness for pickup
- PiVolumeTable: fixed launchpad/SF type IDs from live ESI data (all five
  planet types verified); P0 volume 0.01→0.005 m³; added missing type 2308
- Pull-to-refresh via PullToRefreshBox; PlanetViewModel exposes isRefreshing
- Data freshness label "data Xm ago" per colony (ESI colony TTL ~10 min)
- ColonySummary.dataFetchedAtEpochSeconds captured right after fetchColony()

SecureStorage (Android):
- Catch AEADBadTagException on startup after reinstall — Keystore key is
  invalidated, stale prefs file decryption fails; fix: delete & recreate

Wallet journal:
- Added description field; improved displayName mapping (market_transaction
  shows buy/sell direction, planetary_construction, transaction_tax, AIR)
- Filter JOURNAL_NOISE_TYPES (market_escrow, planetary_construction); show 5

iOS/build warnings:
- @Preview import: org.jetbrains.compose → androidx.compose (5 files)
- shared/build.gradle.kts: -Xexpect-actual-classes compiler flag
- SecureStorage.ios.kt: @OptIn(BetaInteropApi::class) for readBytes()
- Crypto.ios.kt: remove redundant .toInt() on CC_SHA256_DIGEST_LENGTH

+++ shared/src/iosMain/kotlin/com/podforeve/tracker/platform/Crypto.ios.kt
@@ -12,7 +12,7 @@ import platform.Security.kSecRandomDefault
 
 @OptIn(ExperimentalForeignApi::class)
 actual fun sha256(data: ByteArray): ByteArray {
-    val output = ByteArray(CC_SHA256_DIGEST_LENGTH.toInt())
+    val output = ByteArray(CC_SHA256_DIGEST_LENGTH)
     data.usePinned { inputPinned ->
         output.usePinned { outputPinned ->
             CC_SHA256(inputPinned.addressOf(0), data.size.toUInt(), outputPinned.addressOf(0).reinterpret())
